requirements.txt
requests==2.13.0
pycrypto==2.6.1